+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 * Add support for event notifications
 * Add more Scheduler support
+* Add metadata support for `Calendar`, `Message` and `Account`
 * Modify `exchange_code_for_token` to allow returning a full body
 
 ### 5.5.0 / 2021-10-28